requirements.txt
openai-whisper
torch
transformers
sacremoses
ffmpeg-python
tqdm
opencv-python
pillow
requests
openai
anthropic 